Move Feasibility into core; reuse the existing infeasibility renderer
The Feasibility helper had no JUnit coupling — it consumed only Criterion,
BaselineProvider, TestIntent, and FactorBundle, all in punit-api/punit-core.
Reviewer was right: this is a core concept layered just above statistics,
not a JUnit-extension concern. Moved to org.javai.punit.engine.criteria
alongside BernoulliPassRate (the only criterion the gate acts on).

Drops the local formatInfeasibilityDiagnostic that duplicated the existing
InfeasibilityMessageRenderer in core. The legacy ProbabilisticTestExtension
was already consuming that renderer for the same gate; now both pipelines
share one source of wording. The renderer itself moves from ptest.engine
(legacy extension's neighborhood) to org.javai.punit.reporting (where the
other formatting helpers live), and is made public so engine.criteria can
reach it without depending on the legacy package.

Net effect: math in statistics, prose in reporting, gate in engine.criteria,
JUnit wiring stays in junit5. Each layer one step above the one below.
